+++ my-vue3-app/server/app cpoy.js
@@ -1,0 +1,45 @@
+const express = require('express');
+const http = require('http'); // Use 'http' instead of 'https'
+const path = require('path');
+const socketIo = require('socket.io');
+
+const PORT = process.env.PORT || 3010;
+
+const app = express();
+const server = http.createServer(app); // Use HTTP server
+
+const io = socketIo(server, {
+  cors: {
+    origin: [`http://localhost:${PORT}/`,"http://hyblab.polytech.univ-nantes.fr","https://hyblab.polytech.univ-nantes.fr",
+            `http//hyblab.polytech.univ-nantes.fr:${PORT}/`,"http://192.168.1.153",`http://192.168.1.153:${PORT}/`,
+            `https//hyblab.polytech.univ-nantes.fr:${PORT}/`,"http://hyblab.polytech.univ-nantes.fr/ocean-2/",
+            "https://hyblab.polytech.univ-nantes.fr/ocean-2",],
+    methods: ["GET", "POST"]
+  }
+});
+
+const addr_server = server.listen(PORT, function () {
+    const host = addr_server.address().address;
+    const port = addr_server.address().port;
+
+    console.log("Server running at http://%s:%s", host, port); // Change to 'http'
+});
+
+app.use('/ocean-2', express.static(path.join(__dirname, '../dist')));
+
+app.get('/ocean-2/userpage', (req, res) => {
+  res.sendFile(path.join(__dirname, '../dist', 'user.html'));
+});
+
+app.get('/ocean-2/datapage', (req, res) => {
+  res.sendFile(path.join(__dirname, '../dist', 'data.html'));
+});
+
+
+io.on('connection', (socket) => {
+  console.log('A user connected');
+
+  socket.on('send message', (msg) => {
+    io.emit('new message', msg);
+  });
+});
